chore: user microservice app 생성
