+++ apps/user/src/main.ts
@@ -1,0 +1,8 @@
+import { NestFactory } from '@nestjs/core';
+import { UserModule } from './user.module';
+
+async function bootstrap() {
+  const app = await NestFactory.create(UserModule);
+  await app.listen(process.env.port ?? 3000);
+}
+bootstrap();
