+++ apps/user/src/user.controller.spec.ts
@@ -1,0 +1,22 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { UserController } from './user.controller';
+import { UserService } from './user.service';
+
+describe('UserController', () => {
+  let userController: UserController;
+
+  beforeEach(async () => {
+    const app: TestingModule = await Test.createTestingModule({
+      controllers: [UserController],
+      providers: [UserService],
+    }).compile();
+
+    userController = app.get<UserController>(UserController);
+  });
+
+  describe('root', () => {
+    it('should return "Hello World!"', () => {
+      expect(userController.getHello()).toBe('Hello World!');
+    });
+  });
+});
