+++ apps/user/src/user.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common';
+import { UserService } from './user.service';
+
+@Controller()
+export class UserController {
+  constructor(private readonly userService: UserService) {}
+
+  @Get()
+  getHello(): string {
+    return this.userService.getHello();
+  }
+}
