+++ apps/user/src/user.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { UserController } from './user.controller';
+import { UserService } from './user.service';
+
+@Module({
+  imports: [],
+  controllers: [UserController],
+  providers: [UserService],
+})
+export class UserModule {}
