+++ apps/user/src/user.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class AppService {
+export class UserService {
   getHello(): string {
     return 'Hello World!';
   }
